api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.28.1
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
SQLAlchemy==2.0.32
langchain==1.2.13
langchain-openai==1.1.12
pgvector==0.3.6
psycopg[binary]==3.2.12
